refactor(web): extract pages-table component into a folder structure
